feat(depends_linux): обновил установку RMarkdown и Latex

+++ tools/depends_linux.sh
@@ -16,32 +16,39 @@ echo ""
 echo "~~~~~~~~~~~~~~~~~~~"
 echo "Установка RMarkdown"
 echo "~~~~~~~~~~~~~~~~~~~"
-Rscript -e "install.packages('rmarkdown', repos='http://cran.us.r-project.org')"
-Rscript -e "install.packages('reticulate', repos='http://cran.us.r-project.org')"
-Rscript -e "reticulate::virtualenv_remove()"
-Rscript -e "reticulate::virtualenv_create()"
+if ! [ $(Rscript -e "system.file(package='rmarkdown')") ]; then
+    Rscript -e "install.packages('rmarkdown', repos='http://cran.us.r-project.org')"
+    Rscript -e "install.packages('reticulate', repos='http://cran.us.r-project.org')"
+    Rscript -e "reticulate::virtualenv_remove()"
+    Rscript -e "reticulate::virtualenv_create()"
+fi
 
 echo ""
 echo "~~~~~~~~~~~~~~~"
 echo "Установка Latex"
 echo "~~~~~~~~~~~~~~~"
-wget https://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
-zcat < install-tl-unx.tar.gz | tar xf -
-rm -rf install-tl-unx.tar.gz*
-cd install-tl-*
-perl ./install-tl --no-interaction --scheme scheme-medium
-tlmgr install tcolorbox hyperref amsmath setspace indentfirst array
+if ! [ "/usr/bin/latex" ]; then
+    wget https://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
+    zcat < install-tl-unx.tar.gz | tar xf -
+    rm -rf install-tl-unx.tar.gz*
+    cd install-tl-*
+    sudo perl ./install-tl --no-interaction --scheme scheme-medium
+    sudo tlmgr install tcolorbox hyperref amsmath setspace indentfirst array
+    rm -rf install-tl-*
+fi
 
 
 echo ""
 echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
 echo "Установка AnonymicePro Nerd Font"
 echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
-wget https://github.com/ryanoasis/nerd-fonts/releases/download/v3.1.1/AnonymousPro.zip
-unzip AnonymousPro.zip -d fonts
-rm -rf AnonymousPro.zip
-cd fonts
-cp *.ttf /usr/share/fonts
+if ! [ $(fc-list | grep "Anonymice") ]; then
+    wget https://github.com/ryanoasis/nerd-fonts/releases/download/v3.1.1/AnonymousPro.zip
+    unzip AnonymousPro.zip -d fonts
+    rm -rf AnonymousPro.zip
+    cd fonts
+    sudo cp *.ttf /usr/share/fonts
+fi
 
 echo ""
 echo "~~~~~~~~~~~"
